requirements.txt
streamlit
streamlit-drawable-canvas
torch
torchvision
numpy
pillow
joblib
python-multipart